NavDrawer skeleton (stubs)
Added stubs to these classes. MainNavDrawer is a concrete implementation
of NavDrawer.
NavDrawer declares and implements the following inner classes:
NavDrawerItem (Generic item for the NavDrawer), BasicNavDrawerItem (a
concrete implementation of a NavDrawerItem, shows text, a badge and an
icon), and ActivityNavDrawerItem (concrete implementation of a
NavDrawerItem that starts other Activities).